refactor: PA Data Edit Company css
