+++ src/components/PromotionAdmin/DataEdit/Company/InfoCard/Basic.tsx
@@ -1,0 +1,74 @@
+import { getCompanyData } from '@/apis/PromotionAdmin/dataEdit';
+import { ICompanyData } from '@/types/PromotionAdmin/dataEdit';
+import React from 'react';
+import { useQuery } from 'react-query';
+
+import { Wrapper, ContentBlock, InputWrapper, InputTitle } from '../CompanyFormStyleComponents';
+
+import styled from 'styled-components';
+import { DATAEDIT_TITLES_COMPONENTS } from '../StyleComponents';
+
+const Basic = () => {
+  const { data, isLoading, error } = useQuery<ICompanyData, Error>(['company', 'id'], getCompanyData);
+
+  if (isLoading) return <>is Loading..</>;
+  if (error) return <>{error.message}</>;
+  return (
+    <Wrapper>
+      {data && (
+        <>
+          <ContentBlock>
+            {DATAEDIT_TITLES_COMPONENTS.Basic}
+
+            <InputWrapper>
+              <InputTitle>
+                <p>Address</p>
+              </InputTitle>
+              <Content>{data.address}</Content>
+
+              <InputTitle>
+                <p>English Address</p>
+              </InputTitle>
+              <Content>{data.addressEnglish}</Content>
+
+              <InputTitle>
+                <p>Phone Number</p>
+              </InputTitle>
+
+              <Content>{data.phone}</Content>
+
+              <InputTitle>
+                <p>Fax Number</p>
+              </InputTitle>
+
+              <Content>{data.fax}</Content>
+            </InputWrapper>
+          </ContentBlock>
+        </>
+      )}
+    </Wrapper>
+  );
+};
+
+export default Basic;
+
+const Content = styled.div`
+  display: flex;
+  align-items: center;
+  padding: 12px;
+  font-size: 14px;
+  font-family: ${(props) => props.theme.font.regular};
+  min-height: 15px;
+  box-shadow: 1px 1px 4px 0.1px #c6c6c6;
+
+  & p,
+  span {
+    font-size: 14px;
+    font-family: ${(props) => props.theme.font.regular};
+    line-height: 20px;
+  }
+`;
+
+const TitleWrapper = styled.div`
+  display: flex;
+`;
